pyproject.toml
[project]
name = "text-to-sql-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.25.0",
    "google-generativeai>=0.8.5",
    "instructor>=1.9.2",
    "ipython>=9.4.0",
    "pandas>=2.3.1",
    "pydantic>=2.11.7",
    "streamlit>=1.47.0",
]
